refactor: 大規模リファクタリング - 重複コード削除と単一責任原則適用
**コンポーネント共通化:**
- StatCard を独立コンポーネントとして抽出
- RecipeRowView を共通コンポーネント化
- RecipeFormView で新規作成・編集を統合

**重複コード削除:**
- RecipeListView から 250行以上の重複フォーム削除
- SimpleHomeView から StatCard 重複実装削除
- 複数の Add/Edit フォームを統合

**バッジシステム完全削除:**
- BadgeViews.swift, BadgeSystem.swift, Badge+Extensions.swift 削除
- HistoryStatsView からバッジタブ削除
- 関連する UI コンポーネント削除

**ファイル行数最適化:**
- RecipeListView: 1050行 → 793行 (25%削減)
- HistoryStatsView: 711行 → 692行 (3%削減)
- SimpleHomeView: 697行 → 655行 (6%削除)

**レガシーファイル削除:**
- 未使用の StatsView.swift, HomeView.swift, AddRecipeFormView.swift
- 重複する ContentView.swift

**アーキテクチャ改善:**
- 単一責任原則の適用
- コンポーネントの再利用性向上
- 保守性の向上
