+++ CookTrackerProject/CookTracker/CookTracker/Views/SimpleHomeView.swift
@@ -124,7 +124,7 @@ struct SimpleHomeView: View {
         .navigationTitle("CookTracker")
         .navigationBarTitleDisplayMode(.large)
         .sheet(isPresented: $isShowingAddRecipe) {
-            CoreDataAddRecipeView()
+            RecipeFormView()
         }
         .sheet(isPresented: $isShowingTimer) {
             CookingTimerView(timer: sessionManager.sharedHelperTimer)
@@ -644,48 +644,6 @@ struct AddRecipeSheetView: View {
     }
 }
 
-// MARK: - Stat Card Component
-struct StatCard: View {
-    let icon: String
-    let iconColor: Color
-    let title: String
-    let value: String
-    let subtitle: String
-    
-    var body: some View {
-        VStack(spacing: 8) {
-            // アイコン
-            Image(systemName: icon)
-                .font(.title2)
-                .foregroundColor(iconColor)
-            
-            // タイトル
-            Text(title)
-                .font(.caption)
-                .foregroundColor(.secondary)
-                .fontWeight(.medium)
-            
-            // 値
-            Text(value)
-                .font(.title3)
-                .fontWeight(.bold)
-                .foregroundColor(.primary)
-            
-            // サブタイトル
-            Text(subtitle)
-                .font(.caption2)
-                .foregroundColor(.secondary)
-                .multilineTextAlignment(.center)
-        }
-        .frame(maxWidth: .infinity)
-        .padding(.vertical, 16)
-        .padding(.horizontal, 8)
-        .background(
-            RoundedRectangle(cornerRadius: 12)
-                .fill(Color(.systemGray6))
-        )
-    }
-}
 
 // MARK: - Preview
 struct SimpleHomeView_Previews: PreviewProvider {
